only owners of rooms are allowed to edit and delete, none logged users are redirected to login page

+++ studybud/base/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import render, redirect
 from django.contrib import messages
 from django.db.models import Q
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from .models import Room, Topic
@@ -32,6 +34,10 @@ def loginPage(request):
     context = {}
     return render(request, 'base/login_register.html', context)
 
+def logoutUser(request):
+     logout(request)
+     return redirect('home')
+
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else  ''
 
@@ -50,6 +56,7 @@ def room(request,pk):
     context = {'room':room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -62,9 +69,13 @@ def createRoom(request):
     context = {'form':form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!!')
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -75,8 +86,13 @@ def updateRoom(request, pk):
     context = {'form':form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!!')
+
     if request.method == "POST":
         room.delete()
         return redirect('home')
